requirements.txt
pyserial==3.1.1
requests==2.11.1
